beta release of ui

+++ mainUi.py
@@ -18,46 +18,51 @@ def printBanner():
 def printHelp():
 	print """\nNAME:\n\tPGP2IMG - A picture is worth a thousand encrypted words\n\nVERSION:\n\t0.1\n\nCOMMANDS:\n\tencrypt\t\tEncrypt messages in messages for keybase users\n\t\t\t[Keybase user] [Message] [File path]\n\n\tdecrypt\t\tDecrypt hidden message inside images\n\t\t\t[File path]\n"""
 
-def encryptMessage():
+def encryptMessage(recipient, message, transportFile):
 	printBanner()
-	message = os.system("keybase pgp encrypt " + sys.argv[3] + " -m \"" + sys.argv[4] +"\" | tr '[A-Z]' '[X-ZA-W]' >> " + str(sys.argv[5]))
+	message = os.system("keybase pgp encrypt " + recipient + " -m \"" + message +"\" | tr '[A-Z]' '[X-ZA-W]' >> " + str(transportFile))
 	print "message successfully encrypted"
 
-def encryptFile():
+def encryptFile(recipient, payloadFile, transportFile):
 	printBanner()
-	message = os.system("keybase encrypt " + sys.argv[3] + " -i \"" + sys.argv[4] +"\" >> " + str(sys.argv[5]))
-	os.system("echo \"" + sys.argv[4].upper() +"\" >> " + sys.argv[5])
+	print "recepient = "+ recipient + " payloadFile = " + payloadFile + "transportFile = " + transportFile
+	message = os.system("keybase encrypt " + recipient + " -i \"" + payloadFile +"\" >> " + str(transportFile))
+	os.system("echo \"" + payloadFile.upper().split('/')[-1] +"\" >> " + transportFile)
 	print "file successfully encrypted"
-
-def getEncryptedFile(content):
-	regex= re.compile('BEGIN KEYBASE SALTPACK ENCRYPTED MESSAGE[.][.\S\s]{1,}END KEYBASE SALTPACK ENCRYPTED MESSAGE[.]')
-	match = regex.findall(content)
-	return match
-
-def decryptFile(match):
-	regexName= re.compile('BEGIN KEYBASE SALTPACK ENCRYPTED MESSAGE[.][.\S\s]{1,}END KEYBASE SALTPACK ENCRYPTED MESSAGE[.]\n(.*)')
-	matchName = regexName.findall(content)
-	print matchName
-	match[0] +="\n"
-	#print match[0]
-	tempFile = open("temp.txt","w")
-	tempFile.write(match[0])
-	tempFile.close()
-	print "cat temp.txt | keybase decrypt -o "+ matchName[0]
-	os.system("cat temp.txt | keybase decrypt -o "+ matchName[0])
-	os.remove("temp.txt")
 
 def getEncryptedMessage(content):
 	regex= re.compile('-----YBDFK MDM JBPPXDB-----[.\S\s]{1,}-----BKA MDM JBPPXDB-----')
 	match = regex.findall(content)
 	return match
 
-def decryptMessage(match):
-	os.system("echo \""+ match[0] + "\" | tr '[X-ZA-W]' '[A-Z]' > tempyMcTempFace.txt")
-	pgp = subprocess.check_output(['cat', 'tempyMcTempFace.txt'])
-	pgp = pgp
-	print "Encrypted message: " + subprocess.check_output(['keybase', 'pgp', 'decrypt', '-m',pgp])
+def decryptMessage(transportFile):
+	content = subprocess.check_output(['cat', transportFile])
+	match = getEncryptedMessage(content)
+	if len(match) != 0:
+		os.system("echo \""+ match[0] + "\" | tr '[X-ZA-W]' '[A-Z]' > tempyMcTempFace.txt")
+		pgp = subprocess.check_output(['cat', 'tempyMcTempFace.txt'])
+		print "Encrypted message: " + subprocess.check_output(['keybase', 'pgp', 'decrypt', '-m',pgp])
+	else:
+		print "no hidden message found"
 
+def decryptFile(transportFile):
+	content = subprocess.check_output(['cat', transportFile])
+	regex= re.compile('BEGIN KEYBASE SALTPACK ENCRYPTED MESSAGE[.][.\S\s]{1,}END KEYBASE SALTPACK ENCRYPTED MESSAGE[.]')
+	match = regex.findall(content)
+	if(len(match)!=0):
+		regexName= re.compile('BEGIN KEYBASE SALTPACK ENCRYPTED MESSAGE[.][.\S\s]{1,}END KEYBASE SALTPACK ENCRYPTED MESSAGE[.]\n(.*)')
+		matchName = regexName.findall(content)
+		print matchName
+		match[0] +="\n"	
+		tempFile = open("temp.txt","w")
+		tempFile.write(match[0])
+		tempFile.close()
+		print "cat temp.txt | keybase decrypt -o "+ matchName[0]
+		os.system("cat temp.txt | keybase decrypt -o "+ matchName[0])
+		os.remove("temp.txt")
+	else: 
+		print "no hidden file found"
+'''
 #user selects option to encrypt a file.
 if len(sys.argv) == 6 and sys.argv[1] == "encrypt" and sys.argv[2] == "-i":
 	if os.path.isfile(sys.argv[4]) and os.path.isfile(sys.argv[5]):
@@ -76,15 +81,14 @@ elif len(sys.argv) == 6 and sys.argv[1] == "encrypt" and sys.argv[2] == "-m":
 elif len(sys.argv) == 3 and sys.argv[1] == "decrypt":
 	printBanner()
 	if os.path.isfile(sys.argv[2]):
-		content = subprocess.check_output(['cat', sys.argv[2]])
-		match = getEncryptedFile(content)
+
 		if (len(match)!=0):
-			decryptFile(match);
+			decryptMessage(match);
 		else:
 			content = subprocess.check_output(['cat', sys.argv[2]])
-			match = getEncryptedMessage(content)
+			match = getEcryptedFile(content)
 			if(len(match)!=0):
-				decryptMessage	(match)
+				decryptFile(match)
 			else:
 				print "no hidden message found"
 	else:
@@ -93,4 +97,4 @@ elif len(sys.argv) == 1 or sys.argv[1] == "help":
 	printHelp()
 else:
 	print "INVALID ARGUMENTS"
-	printHelp()
+	printHelp()'''
